refactor: Reorganize type imports for better maintainability
