+++ src/components/common/display/__tests__/Image.test.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import Image from '../Image';
