+++ src/components/common/display/__tests__/LabelValue.test.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
 import LabelValue from '../LabelValue';
