+++ src/components/common/forms/__tests__/GroupCreationForm.test.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import GroupCreationForm from '../GroupCreationForm';
